make tie_breaking_argmax more secure

+++ src/utils/utils.py
@@ -2,4 +2,4 @@ import numpy as np
 
 
 def tie_breaking_argmax(a: np.ndarray, eps: float = 1e-8) -> int:
-    return np.argmax(a + np.random.random(a.shape) * eps)
+    return int(np.argmax(a + np.random.random(a.shape) * eps))
